+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 ### Cleanup
  * Removed the deprecated crypto.NewSHA256Signer function.
  * Finish removing the `LogMetadata.GetUnsequencedCounts()` method.
- * Removed `TrillianLog.GetLeavesByHash` API.
+ * Removed `TrillianLog.GetLeavesByHash` and `TrillianLog.GetLeavesByIndex`
+   API.
 
 ### Storage refactoring
  * `NodeReader.GetMerkleNodes` does not accept revisions anymore. The